requirements.txt
aiogram>=3.8.0
python-dotenv>=1.0.1
yt-dlp>=2025.0.0
aiofiles>=24.1.0
